Fix Hazel builds on Windows with GHC < 8.6
With GHC < 8.6, long package names would trigger failures with Hazel on
Windows due to path length exceeding the 260 characters limit. While
disabling this limit globally in the Windows 10 registry is possible, this
would not be honored by GHC binaries (such as ghc-pkg), as a
special manifest must be provided for this to work with mingw-derived
executables. This would cause issues in particular with packages compiled
via Hazel, as those are fetched in external repositories, which makes for
even longer file names than the usual.

This commit fixes this issue by using a .bat script on Windows to find out
the "short path" of the package database, and passing that to `ghc-pkg
recache`.

Additionally, we also expose the rts-headers in GHC bindists, as required by
the cbits targets generated by Hazel.