Allow conditional compiler_flags
Prior to this commit it is not possible to specify condition (in
particular per-platform) compiler flags. This is a problem with flags
like `-split-sections` which are useful on Linux and error out on
Darwin.

This commit modifies `register_ghc_nixpkgs` to add a field
`compiler_flags_select` which is essentially fed to a `select()` call in
the BUILD file.