Prefix -ghci-script by exec root
Before the GHCi wrapper located the ghci script relative to the working
directory, which was expected to be the workspace directory. This worked
fine for executing the `@repl` files using `bazel run` as these were
executed from the repository root and not from a dedicated exec root.

In case of a dedicated repl target this will no longer work, as the
target will be executed in its exec root. This change modifies the
wrapper to first change into the workspace directory, and to refer to
the ghci script relative to the exec root. The same is done for library
dependencies.