Revert "Fix Hazel builds on Windows with GHC < 8.6"
This reverts commit 8a6ca46b41cf01e1c10300cfa93073c7965f5cfb.

This hack is detrimental to remote caching on Windows, as it introduces absolute file paths into package databases.
A better solution is to enable long file names on Windows, which is a one-time operation.

An unfortunate consequence is that rules_haskell won't work on Windows for any GHC version prior to 8.6, nor on any Windows version but 10+ ; but any reasonably-sized
build would already fail there because if inherent filesystem limitations of older MINGW binaries, especially in conjuction with Hazel.