start script: A "hello world" can be built & run, documentation
The user is now immediately able to build a simple Haskell binary.
We provide a full minimal example of a working project, getting the
user up & running from a state they know is working (or at least one
we guarantee to work).

Each of the generated bazel files is documented, to ease understanding
of the template.