Set -Wmissing-home-modules to flag potential issues when build is not sandboxed.
With sandboxing turned on, it should never happen that GHC will
accidentally use a source file that wasn't declared as an input to
Bazel. But if sandboxing is turned off, this can happen. And when it
does, surprising things happen. So like Cabal, we ask GHC to warn us
about this situation.

Closes #128.