Hazel: Make excluded C libraries configurable
Some packages list libraries in the extra-libraries section of their
cabal file, which are always provided by the system and do not need to
be provided as Bazel targets. E.g. libstdc++, or libpthread.

Previously, Hazel held a hard-coded list of such libraries for which it
would not expect a user-provided Bazel target.

This change allows users to configure this list of system libraries
themselves by passing empty string entries in the extra_libs argument to
hazel_repositories.

Addresses the user configurability aspect of #834.