ls_modules is more robust with duplicate packages
ls_modules was crashing with an error if an haskell package was
duplicated in the package database. I.e if it exists a least two
package with the same package name and a different package id.

This commit changes the error behavior, it now only crash when the
package is explicitly used by the reexported module phase. So as long
as the duplicate is not explicitly used, it does not crash.

Closes #839