Make providers.bzl public.
Providers are everything but private. They should be considered an
integral part of the interface of a rule. Any downstream dependencies
are free to query the providers passed to them and poke at each of
their fields. We should carefully consider any changes to providers,
because changing providers means changing the interface of a rule.