Don't expose exposed_modules_file in providers
The only rule that was using this was the `haskell_doctest` rule. But
arguably this was an incorrect usage of this file. The reason is as
follows. `haskell_doctest` uses GHCi under the hood. Unlike GHC, GHCi
does *not* use module names to correctly locate source files, unless
the source file paths can be derived from the module name. So if we
have `Foo.hs` defining `Bar`, we should *not* tell `doctest` to load
`Bar`, but instead tell it about `Foo`.

For this reason it would be better for `haskell_doctest` to only talk
about source files and not about modules, i.e. deprecate the `modules`
attribute. But that's a matter for another day.