Split haskell_toolchain
We initially modeled `haskell_toolchain` on `go_toolchain` in
rules_go. `haskell_toolchain` is a macro that defines a private
toolchain implementation rule and then the `toolchain` rule to specify
execution and target constraints. This design, while convenient,
creates problems. Ideally, we want the toolchain implementation
definition to be next to the code, and we want the `toolchain`
metadata to be in a separate repository. That way Bazel can see the
`toolchain` definition without downloading the toolchain itself, and
decide based on the platform constraints whether the toolchain should
be selected during toolchain resolution.

Keeping `haskell_toolchain` and `toolchain` separate matches the
design of the CC rules, where there is likewise a `cc_toolchain` and
`toolchain`.

This is a breaking change: using `haskell_toolchain` now requires
calling it and then also `toolchain`.