Add haskell_cabal_library
This is an alternative solution for handling Cabal libraries published
on Hackage to both Hazel and Nixpkgs. The idea in this one is to use
Cabal to build Cabal packages, and Stack to download source
distributions from a Stackage snapshot and generate a dependency
graph. See #874 for the rationale. The Stack part will be submitted as
the next PR in the series.

This rule exposes all the same providers as `haskell_library`, so one
can be substituted for the other.

This rule is able to build the close to 200 dependencies of Stack. The
vast majority of Cabal packages in Stackage build out-of-the-box,
including those that have system dependencies like `zlib`. Only
packages that do very funky things like attempt creating their own
files in the CWD fail (like Stack itself).