Remove "new-style" haskell_import and related code
This mechanism was introduced in #442 and was meant to be a stepping
stone towards removing the special treatment of prebuilt dependencies.
It was plagued by performance problems related to Nix and never saw
adoption beyond this project's test suite. With the advent of
`stack_install` and `haskell_cabal_library`, we can finally get rid of
it.