Add package name export aliases to nixpkgs toolchain
Haskell packages exported by nixpkgs toolchain can lead to conflict in
bazel names if the environment contains many packages with the same
package name.

This commit introduces a new aliases attribute to the toolchain
registration rule which allows the override of package exported names.

Packages id are matched using a prefix, this way the author does not
need to provide a fully qualified package id, hence this is more
robust to underlying toolchain changes (such as nixpkgs upgrades). For
example, users can only match on package name + version if the
conflict appears between two packages of different version, such as
`ACME-0.0.0.0` and `ACME-0.0.0.1`.

This closes #885 and depends on #886