Separate haskell_import interface-dirs/files
The interface_dirs property on the HaskellInfo provider is only used as
an action input, never to point to the directory path. However, treating
it as a directory carries the risk of overlap with other files, such as
header files or library files. Hence, we add a separate interface_files
attribute to the haskell_import rule which carries the actual interface
files. These are then added to the HaskellInfo provider, while the
interface_dirs attribute is only used to generate the package
configuration file.