Rename stack_install to stack_snapshot
We originally chose `stack_install` to follow `mvn_install` from
rules_jvm_external. But the latter isn't relevant precedent since
`mvn_install` is named that way because `mvn install` is a command
that does morally the same thing. `stack install` does exist but
doesn't have the same meaning. In fact we may introduce
`stack_install` in the future, but then it would mean building
binaries and exposing them to other targets.

We don't call this a Stackage snapshot because any resolver supported
by Stack is allowed. Including custom snapshots in the future,
potentially.