Add a tools argument to Stack and Cabal rules
Tools are built using the host configuration. Currently, they are not
very useful unless they are tools part of a whitelist defined by
Cabal. For example, Cabal lets you customize which happy or c2hs you
want to use. We support that.

Packages that correspond to the whitelisted tools are detected as such
and filtered from the dependency graph in `stackage_snapshot`. That
way only packages that produce libraries remain.