Expand location and make variables in compiler flags
The documentation for both the C/C++ rules and the Java rules
carefully notes in any attribute that is a list of flags to something
that this attribute is subject to "Make variable substitution" (and
bourne shell tokenization). I don't know how to do the latter, and
I don't know that it matters. But for consistency with the Java and
C/C++ rules, we now do the same too.

This can occasionally come in handy. For example to pass in the
location of a tool that belongs to the java toolchain.

[skip ci]