pkgdb_to_bzl now fills the haddock fields
The haddock fields of haskell_import are not read from the global
package database.

In the case of nixpkgs, the haddock path are outside of the WORKSPACE
containing the ghc toolchain. We introduce a path rewriting scheme
which generate local bazel names inside the pkgdb_to_bzl.py file and
let the bazel process create the associated links.