Remove haskell_lint
It is seldom or never used, probably because it's clunky to use. It
had a design flaw: the rule should have been a test rule, when it was
not. Removing it will make way for a `haskell_lint` with a different
purpose: actual lint checking rather than compiler warnings. Users can
continue doing what they already do: set `-Wall -Werror` in the
toolchain `compiler_flags`.