Allow arbitrary Cabal tools
Cabal has a whitelist of tools, for which we can supply a path
explicitly using `--with-PROG`, where `PROG` is the name of the tool.
But in practice `Setup.hs` scripts and the build process in general
sometimes use other tools that are not part of the whitelist. For
example, `postgresql-libpq` calls `pg_config` during configuration.
Unlike the whitelist, these tools can only be supplied by ensuring
they are reachable from the `PATH`. So we extend the `PATH` to make
this true, hoping that not too much else is exposed as a side effect.

We could avoid undeclared inputs leaking, but

1. we can't do this without a hefty performance cost on Windows (no
   symlinks),
2. in practice this should be the responsibility of the workspace
   rules for each dependency to expose just what we want,
3. Cabal rules are not hermetic to begin with, because they need all
   manner of standard utilities.

Adding a test for this would be awkward, and the feature is at a low
risk of regression, so I didn't write one. Testing would require
making all of `@stackage`, and hence most tests, depend on Nixpkgs. We
don't want that, because of bindist builds.