Change the workspace name
This is a breaking change, because macros often generate BUILD files
that assume the repository has a certain name in the WORKSPACE file.
But it's also going to make importing haskell rules BUILD file after
BUILD file a fair bit more pleasant.

The new name is to be consistent with the Bazel rule deployment
guidelines. The guidelines about naming have changed. See
https://docs.bazel.build/versions/master/skylark/deploying.html.