Download Stack if possible
If we don't find a recent enough Stack on the PATH, then just download
one locally and use that. Since this is happening in a workspace rule,
we have no choice but to use prebuilt binaries. This behaviour brings
us closer to what rules_jvm_external does, which likewise downloads
a tool used to compute the dependency graph.

NixOS users will of course need to continue providing Stack in the
shell, because the Stack bindists don't work there, or using the new
`stack` attribute.