Windows: cc_wrapper in bash
The python cc_wrapper is too slow on Windows. See
https://github.com/tweag/rules_haskell/pull/1002

Rewriting it in C++ would be a lot of work as we cannot rely on C++17's
std::filesystem, yet. Including a C++ filesystem library would impose a
large dependency. Fortunately, the Windows cc_wrapper only needs to
shorten library paths, not handle dynamic libraries. It is easier to use
a dedicated bash cc_wrapper on Windows instead.