Don't expose transitive html dirs in HaskellHaddock action
It is common (in particular in bindists) that the .haddock file is in
the html dir. If we pass both as inputs, that confuses the sandbox
setup code, which fails with a "File exists" I/O exception, because
the one input is a prefix of the other. However, the HTML files are
immaterial to Haddock. It just needs to read the interface files. So
we don't expose them.