Make compiler version check happen earlier
We use to have an action that all other actions depended on
transitively: a compiler version check. We remove this check, which is
redundant in the case of bindists, and happens too late in the case of
Nixpkgs. We replace it with a new check that happens at workspace rule
evaluation time, and which is Nixpkgs specific.

Fixes #945