CI: collapse consecutive `nix-shell --run` calls
Calling `nix-shell` has a few seconds overhead (because it has to
re-evaluate `shell.nix` every time it is called).

The string given to `--run` is just inserted into the rcfile bash is
started with, so we can give it a multi-line string and it will
execute it like a bash script.

This way we only pay the `nix-shell` overhead once per CI job.