Use the Cabal name for module reexports
We have two choices for the name of this field: follow Bazel
convention or follow Cabal convention. We would normally do the
former, but I would argue that there is no Bazel precedent in this
case. For package reexports, we have the `exports` field like in Java.
We could use the same precedent for module reexports, but for the
following qualitatitve differences:

* module reexports don't have the same type. We have to map labels for
  packages to reexport specs.
* Module reexports aren't about targets per say, but rather parts of
  a target (aka a "package" in Cabal parlance). The rules don't really
  know about modules, only the compiler does.

So for lack of established Bazel precedent, we follow Cabal precedent.
This has the advantage that attributes that handle modules are all
named the same way (`hidden_modules`, `reexported_modules`, etc).