Test start script from a constant directory
This leads to more cache hits. This is necessary in particular for the
start script, which we no longer checkout in a random folder every
time. Since the folder path is part of the key in the Bazel cache (see
https://docs.bazel.build/versions/master/output_directories.html), the
only way to get cache hits every time is to use the same folder each
time.

Incidentally, this should also solve runaway cache growth.