Let haskell_cabal_library names be arbitrary
By default, the Cabal package name and the rule name are the same. But
using the `package_name` attribute, one can make the rule name and the
Cabal package name differ. We also set the version in a separate
attribute, rather than demand that the version be part of the rule
name.