Handle empty libraries on Hackage
Empty libraries are libraries that don't expose any modules. It is
legal for a Cabal package to have an empty library. We now handle this
case properly. However, it is hard to do so in full generality. Since
these packages are rare and old, we use a blacklist for now of known
empty packages and do something special when we encounter a package in
that list.

Fixes #999