Inject vendored packages into stack dependency graph
Before vendored packages were only used when generating the `BUILD` file
to override `deps` attribute to `haskell_cabal_library`. However, this
did not affect stack's dependency graph generation. A vendored package
could have a different Cabal file introducing different package
dependencies. Furthermore, the generated target for a vendored package
would still point to the unvendored original version.

This change takes this into account, by pointing stack to the source
directories of the vendored packages. It also changes the generated
targets to be aliases to the vendored target.

This requires the vendored target to contain a Cabal file at the
top-level which is not a generated file.