Cabal: Unique output directories
Make the output of haskell_cabal_library|binary depend on the target
name. This is to avoid collisions with other targets in the same
package and allow multiple Cabal targetes in the same package.

Try to have the package name appear only once in the output path. This
is to avoid too long file paths, which might lead to issues on Windows.