Check if haddock paths exists
Ghc package database may contain not existing paths, we check that they
really exists before writing them in bazel database.

The logic is a bit convoluted here, because we have first to replace and
check $topdir. Also, if the paths are "local" (i.e. contained in the
current repository), there is no need for symlink. But whatever happen,
we need to check if the file exists.

Closes #1075