glob --> breadth_first_walk
globbing to search for the solib directory can be slow if the working
directory contains a large number of files or directories.

This change uses a breadth first recursive directory walk instead of
`glob`. We are looking for the shortest path matching the solib
directory, the first match in a breadth first search will be the
shortest one.