toolchain: use C.UTF-8 as default locale
`C.UTF-8` is a locale supported by [fairly recent
glibc](https://sourceware.org/glibc/wiki/Proposals/C.UTF-8)
and is the same as the default `C` locale, but with the following
differences:

* LC_IDENTIFICATION: Mentioned for completeness
* LC_COLLATE: Sort using the Unicode codepoint
* LC_CTYPE: UTF-8 encoding

The big advantage this gives us is that we don’t implicitely require a
locale_archive to exist on the system. This is especially handy for
e.g. minimal containers that don’t provide a locale archive (e.g. on
CI).