Set empty values for protobuff HaddockInfo informations
`package_id` cannot be None, because it is used in the rest of the
process. It is now set to the real name of this haskell package.

The protobuf process does not generate haddock informations, as reported
in #1030, leading to missing documentation link and build failures.

`transitive_html` must contain a key with the current package_id. This
is fixed to an empty list.

This is a partial fix for #1030 because it fixs the build failure