Python / Haskell tests: fix the type of imported functions
This fixs a longstanding issue. hs_init was called without arguments,
which is wrong because the right prototype is:

int hs_init(int argc, char** argv)

This was leading to some invalid read in memory and sometimes it was
segfaulting.

I force the type in the prototype of hs_init and during the call to
add_one_hs