Don't propagate the source files of `exports` packages
When a haskell package has some package `exports`, we build a new
`HaskellInfo` by combining this package's own info and all of it
re-exported ones. But we also used to list all the source files of all
the re-exported packages in this new `HaskellInfo`, which doesn't make
sense in general and broke haddock.

Fix this by just putting the source files of the current package in
the `HaskellInfo` provider