haskell/gen_ghc_bindist.py: use python3 in shebang
We use python3 in all other python shebangs in the repo anyways, so it's
reasonable to assume we require python3 to be present.

Debian-style distros didn't yet alias python to python3, but provide it
in the python2 package.

I don't really think we should require Python 2 too, but just Python 3
here aswell.