Support for package database with long package name
This is a fix for #1165

GHC 8.8 changed the output format of `ghc-pkg dump` and introduce line
breaks in `id:` fields.

The parser was simply using `.split(':')` on line containing `id:`. This
no longer work.

This commit introduces a new parser which parses line by line the file
until entering a field description.

The error only appears with GHC 8.8, so I have no reproduction code
available on the current codebase. However the GHC 8.8 branch for
rules_haskell will contain a non regression test for this bug.