link_library: add `compiler_flags`
`compiler_flags` (local or global) can be used by user to override the
link phase done by GHC. The `-threaded` flag is an usuall example of
that and that's why `compiler_flags` was used in `link_binary`.

Until this commit, `link_library` was not taking `compiler_flags` into
account. This commit changes that.

One use case may be to pass flags custom flags to the associated link,
such as `-fsanitize=undefined` which is mandatory if we link with a C
library which uses gcc/clang sanitizers.

-