haskell: don't call cc_toolchain.ar_executable et al
The attributes of cc_toolchain, such as ar_executable and
various others, were anomalously both fields and methods,
so that expressions of the form x.f and x.f() had the same value.
This was a bug in Bazel that will soon be fixed, causing
the x.f() form to no longer be valid.

This change replaces x.f() with x for attributes of toolchain.

Change-Id: I9eb5853c34a90d1c616d4d0c7f7f2bc5b632af79