Fix cyclic aspect dependencies in proto_library
haskell_cc_libraries_aspect depends on the CcInfo and optionally
HaskellInfo providers of a target. In the case of proto_library targets
these providers are returned by the _haskell_proto_aspect. That aspect
in turn requires HaskellCcLibrariesInfo in all its dependencies. Bazel
does not allow this kind of cyclic dependency and one aspect will not be
able to observe the other.

To work around this we instead generate HaskellCcLibrariesInfo within
_haskell_proto_aspect and bundle it in HaskellProtobufInfo.