Define Bazel version in one shared location
Instead of each CI pipeline separately defining the Bazel version, we
define it in one shared location `.ci/bazelversion`. Ideally, we would
use `.bazelversion` in the top-level which would be compatible with
[bazelisk](https://github.com/bazelbuild/bazelisk). Unfortunately, this
is incompatible with the way Bazel is packaged in nixpkgs right now.
See, `.ci/check-bazel-version`.

The pipelines that use nixpkgs to provision Bazel check that the Bazel
version matches using `.ci/check-bazel-version`. The pipelines that
fetch a Bazel binary distribution use `.ci/fetch-bazel-bindist` which in
turn consults the `.ci/bazelversion` file. To avoid issues with
corrupted downloads we check the downloaded bindist against expected
sha256 hashes defined in `.ci/bazel-sha256`.

The files `.ci/bazelversion` and `.ci/bazel-sha256` need to be kept in
sync. The script `.ci/update-bazel-version` helps with this. Given a
Bazel version number it will update both files. The nixpkgs revision
needs to be updated separately. The `MAINTAINERS.md` file has been
updated accordingly.