Escape unknown bytes in package config
This section of code is adjusting the paths in the package config. It
does this by opening a file, `sed`ing some lines, and writing it back.

As this code is run on multiple systems, each system can have a
different default encoding. Some windows systems use CP-1292 as the
default. In which case, a file with UTF-8 encoding will have unreadable
bytes for CP1292. `open` defaults to raising exceptions when there is an
encoding error.

While we could default to assuming every package config file is UTF-8,
there's another option. Since this code exists to `sed` something (and
it doesn't seem to care what the non-regex bytes are), we ought to be
able to escape the unknown bytes to the private use area, do the changes
we want, then unescape them when writing the file back out.

This should allow us to effectively ignore bytes that would normally
fail with the default encoding.

There's more information about the different ways to handle errors under
the `error` section:
https://docs.python.org/3/library/functions.html#open.