Refactor `--config` setup
Define top-level configurations for the user to specify, e.g. `--config
linux-nixpkgs`. These will then define the required flags for that
platform/toolchain combination.

Configure CI by writing `--config` flags to `.bazelrc.local` instead of
passing them on the command-line. This makes sure that the same
configuration is used even if Bazel is called indirectly through a
script. This is useful for tools like `ghcide` and `hie-bios`.