Fix cc_wrapper RUNPATH rewriting
The cc_wrapper removes unneeded RUNPATHs and merges redundant RUNPATHs.
The corresponding logic requires the final output binary to be produced
at a path from which the relative RUNPATH entries to Bazel managed
dynamic libraries are valid. This change accomplishes this by building
within a temporary directory that is a sibling of the final output
directory.