Define setup_deps as string in stack_snapshot
Otherwise, it is not legal to depend on a package from within the same
snapshot. E.g. the package `polysemy` from `@stackage` could not have a
`setup_deps` dependency on `@stackage//:cabal-doctest`, which is very
inconvenint.

The dependency does not have to be resolved at repository rule
evaluation time, only at build time of the Cabal target, hence it is
sufficient to pass the labels as strings.