Add missing library files to the deps of the repl targets
The haskell_repl targets don't directly depend on the `.so` files of
their dependencies, but on the symlink to these .so files that's in the
`_solib_k8` directory.
This works fine locally (because building the symlink requires building
the .so file and the repl run isn't sandboxed), but fails with a remote
cache and `--remote_download_minimal`.