Add the interface files of the deps of the repl targets
The haskell_repl targets don't explicitely depend on the interface files
(`.dyn_hi`) of their dependencies.
This works fine locally (because building the dependencies requires building
the interface files too and the repl run isn't sandboxed), but fails
with a remote cache and `--remote_download_minimal`, so add them to the
dependency tree of the rule.