HaskellInfo: Merge static|dynamic_libraries
The distinction between static and dynamic libraries for linking and C
dependency purposes are tracked in the CcInfo provider instead of the
HaskellInfo provider. Based on that alone the static_libraries and
dynamic_libraries fields could be removed completely. However, GHC has
complicated logic around which versions of a Haskell library it will use
for which steps of the compilation pipeline. In particular, static
libraries are used for the linking of final mostly static artifacts,
dynamic libraries are used for the linking of finaly dynamic artifacts
and for GHCi/Template Haskell (assuming a dynamic RTS), and profiling
static artifacts are used for linking profiling artifacts. Combinations
of the above can be required at the same time. E.g. when building code
using Template Haskell in profiling mode. The HaskellInfo provider's now
merged hs_libraries field is a good place to track libraries for all
these different [GHC "ways"][ghc-ways].

[ghc-ways]: https://gitlab.haskell.org/ghc/ghc/-/wikis/commentary/rts/compiler-ways.