Separate dependency resolution from unpack
Previously, the strategy was to first fetch all packages listed by the
user in the `packages` attribute to `stack_snapshot` and then use them
as "local packages" to determine all transitive dependencies. This
interleaves the fetching of remote packages with the dependency
resolution which prevents us from persisting the results of dependency
resolution in a lock file.

The new strategy implemented by this commit is to instead generate a
local "dummy" package that lists all packages specified by the user as
its dependencies and generate a `stack.yaml` file that captures all
Cabal flags, package version overrides, and vendored packages. Stack is
then invoked on these two files to determine the transitive dependency
graph without first fetching any packages into the external Bazel
workspace. The package sources are only fetched in a later step.

This separation allows to store the package metadata in a lock file and
let Bazel fetch package sources separately based on the information in
the lock file in future.

I have verified that the resulting external workspaces are equivalent
up to expected differences by comparing the results of `bazel fetch
@stackage//:all` in the `example` directory, similarly for `@stackage`
and `@ghcide` in the project directory. Note, dictionaries are iterated
in insertion order. To compare the results before and after this commit
one has to sort the contents of `package.bzl` and `BUILD.bazel` by
package name first.