Add test for Protobuf name clash issue
This indicates the issue highlighted in #1111 does not appear
with the recent proto-lens (e.g. from the version of Stackage pinned
in this commit).

This also shows that Haskell modules generated from Protobuf imports
(such as google.protobuf.FileDescriptorSet) are hidden by default
and introduces a workaround for that issue.