Separate resolving and downloading packages
The goal is to be able to persist the results of resolution in a
lock-file and be able to download and build remote packages solely based
on that lock-file.

This replaces `_compute_dependency_graph` by `_resolve_packages` and
`_download_packages`. Additionally, the parsing of the results of `stack
ls dependencies json` is simplified. `_compute_dependency_graph` did an
unnecessary additional transformation from the `stack` output to
`all_pacakges` for historical reasons. This has been simplified.