Enable pinning with stack_snapshot
Adds an optional attribute `stack_snapshot_json` to `stack_snapshot`
which pins the result of `stack ls dependencies json` and additional
reproducibility information. If specified, then `stack_snapshot` will
not need to invoke `stack` to resolve package versions and dependencies
or to fetch Hackage or archive dependencies.

- This makes re-fetching `stack_snapshot` repositories faster as Hackage
  and archive dependencies can be cached in Bazel's repository cache,
  and `stack` does not need to update it's database to determine package
  versions and dependencies.
- Additionally, this avoids issues due to concurrent invocations of
  `stack` which can fail due to a race condition on the Hackage security
  lock.

The lock file can be generated or updated by invoking the script
`@<workspace>-unpinned//:pin`.

If the `stack_snapshot_json` attribute is not specified, then
`stack_snapshot` will behave as before and resolve and fetch external
packages using `stack` on every Bazel fetch.

The resulting `stack_snapshot` workspaces are identical with the
exception of revisions of Cabal files. `stack ls dependencies json` does
not expose the Hackage revision of any given package used in the
snapshot, only the version. However, Cabal file downloads are not
reproducible as Hackage revisions may change them over time and a
revision >0 may be required for a given snapshot to avoid issues with
version ranges. The pinning implementation pins the current git revision
of the [all-cabal-hashes][all-cabal-hashes] repository and downloads
Cabal files from that revision. This may potentially use a later
revision than used in a particular snapshot.

[all-cabal-hashes]: https://github.com/commercialhaskell/all-cabal-hashes

`stack ls dependencies json` does not provide subdirectory information
of `archive` dependencies. This change includes a heuristic which
expects the Cabal file at the archive root or underneath a top-level
directory and defines `stripPrefix` accordingly.