expand make variables in compiler and REPL flags
The expansion requires the target's rule `ctx`, so we perform expansion
at the target and track the expanded flags in the `HaskellInfo` provider
to make it available to `haskell_repl`.

`user_compiler_flags` is distinct from `compiler_flags` in that the
former holds only flags explicitly provided by the user while the latter
includes flags that are generated by rules_haskell.