haskell_repl: Track data and runfiles explicitly
In case of `haskell_binary` or `haskell_test` targets the
`DefaultInfo.runfiles` will include the binary itself. This meant that
if `collect_data == True` then the REPL depended on the compiled binary
of targets that should be loaded into the REPL by source. I.e. if a
module that should be loaded into the REPL by source cannot be compiled
due to a compiler error then it can also not be loaded into a REPL with
`collect_data = True`. This is very inconvenient as the REPL might help
to fix the compiler error.

This changes fixes this by tracking data dependencies and runfiles
explicitly rather than just relying on `ctx.runfiles(collect_data =
True)`. In particular, in case of targets that should be loaded by
source we will not include the full `DefaultInfo.runfiles` (which would
include the final binary of a `haskell_binary|test`), but instead we
only include the files and runfiles of `data` dependencies.

Closes #1380