Align Bazel.Runfiles to the specification
In particular, the discovery of `.runfiles` or `.runfiles_manifest`
should be based on the value of `argv[0]`. Such that a runfiles tree
placed next to a symbolic link that is invoked is discovered, or such
that an outer binary calling this binary can point to a different
runfiles tree by setting `$RUNFILES_DIR` or `$RUNFILES_MANIFEST_FILE`.