Add regression tests for rlocation overrides
The Haskell runfiles library should match the behaviour of other
runfiles libraries that a wrapping `sh_test` or other outer binary
target can override the runfiles tree of an inner `haskell_binary`
target.

In the provided test-cases the runfiles tree of `hs-runner` should be
overriden by the runfiles trees of the tests `hs-foo` and `hs-bar`.