Fix incorrect path prefixing when linking
When adding support for fully-statically-linked binaries, we taught
`link_libraries` how to prefix its linker directories using
`path_prefix`. At the time it was decided that it would be up to the
caller to add trailing slashes to `path_prefix` as necessary, to keep
`link_libraries` simple. In threading this through the various callers
of `link_libraries`, we introduced a bug whereby the _empty_
`path_prefix` is transformed into `"/"`, which breaks a few things. This
commit thus makes the following changes:

* `link_libraries` _does_ now automatically add a trailing slash to a
  non-empty `path_prefix` argument. This doesn't seem overly restrictive
  and certainly at present we don't use non-directory prefixes as
  arguments to this function.

* The comment that explains why REPL scripts must pass a `path_prefix`
  to deal with the fact that they don't run in the execroot is moved to
  arguably the better position -- `_create_repl` as opposed to
  `_compiler_flags_and_inputs`. `_compiler_flags_and_inputs` is actually
  also called by `_create_hie_bios`, which needs a similar workaround
  but at present doesn't provide one (to be addressed in a future
  commit).