Only pass -no-pie for binary targets
When using GHC 8.10, passing `-no-pie` to the linker when building
libraries leads to obscure linker errors. Since PIE pertains to
executables, it doesn't make much sense passing that flag when
building libraries anyways. Removing it, both for `haskell_library`
and `haskell_cabal_library` fixes GHC 8.10 support. We make sure to
keep the flag for executables. This would otherwise negate the effect
of #1388.