Check each file we might overwrite or clash with.
This change is more explicit in its messaging about each file that exists and that we'd be overwriting. It avoids false positive clashes with directories such as the common `build` directory in many repos.

```
> ./rules-haskell-start
INFO: Creating a WORKSPACE file based on GHC bindists. If you want to use a nix-based setup (e.g. on NixOS), call with `--use-nix`. See `--help` for more info.
ERROR: The current directory already has a .bazelrc file we don't want to overwrite.
```