Example usage for the lcov merger
This is what is currently used on a client codebase.

HPCLcov is a tool which converts hpc ".tix/.mix" reports to the lcov
format. It is based from hcp-lcov hackage package, but proposes an
interface which works with bazel assumptions.

Note that the code can be highly simplified, especially, some options
are now useless.

The shell.nix, wrapper script and .bazelrc works are here to fill
LCOV_MERGER, which is used by bazel as a post-processing for coverage.
Unfortunately, it is not possible to pass a bazel label, hence the need
for outside bazel build, and the environment definitions.