Avoid patching bindist bin symlinks
On MacOS Big Sur it is no longer allowed to apply `sed` in-place changes
on a symlink. This change makes sure that we only change regular files
in the GHC bindist bin directory.

The symlink resolution did not handle relative symlinks correctly. This
was not an issue so far as symlinks were implicitly converted to regular
files.